# Add missing INLINE's on EnumContainers