@ChrisPenner 2024-12-12
From staring at the core I was able to determine that the difference is actually because `CCache` is _not_ getting unboxed on this branch.

Something about inlining these convinced GHC NOT to unbox the CCache, and it improves the speed by not passing the extra dozen args to the `eval` worker.

Dan and I confirmed we get a similar speedup by just removing the `!`s on CCache in eval and exec.

That said, this PR  should probably go in anyways since there's no reason to NOT inline these, and clearly it's helping GHC be at least a little smarter :)